fix: Make adjacent nodes in hex have dist 1
Previously the nodes were separated in the y axis by 1, which results
in adjacent nodes forming isosceles triangles instead of equilaterals.
Scaling the nodes vertically to shrink the distance between nodes
results in adjacent nodes having distance one. This commit contains
one failing test for the purpose of demonstrating the requirement
for re-initialisation.